added game preparer for api only
